requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
langchain
langchain-google-genai
python-dotenv
langchain-groq 
groq

